requirements_enhanced.txt
Flask==2.3.3
WeasyPrint==60.2
gunicorn==21.2.0
reportlab==4.0.4
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2 